♻️ refactor(account): restructure account endpoints and shared resources
-【feat】move `ApiInfo.cs` to `Featrues/Account` directory
-【feat】create `GetDetailsUserEndPoint.cs` in `Featrues/Account`
-【refactor】move shared contracts and tools to `Shared` directory
-【refactor】update namespaces and references to reflect new directory structure
-【refactor】remove unused `DrMeet` namespace
